app/webapp/requirements.txt
bcrypt==3.2.0
psycopg2-binary==2.9.1
Flask==3.0.1
kubernetes